Fix the asynchronous worker timeout issue
Workers silent for more than this many seconds are killed and restarted. By silent, it means silent from the perspective of the arbiter process, which communicates with the workers through a temporary file. If the worker is busy sending data, it does not update that file. From the perspective of the arbiter, the worker is missing heartbeats.

We add sleep(0) between generator iterations to give the worker a break and update the heartbeat file to keep the connection and worker process alive.